feat : 카테고리별 분류 후 각 컴포넌트 생성
